# swagger.yaml
openapi: 3.0.3
info:
  title: Онлайн библиотека песен
  version: 0.0.1
paths:
  /songs:
    get:
      parameters:
        - name: group 
          in: query
          required: true
          schema:
            type: string
        - name: song
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: size
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoupletsDetails'
        '400':
          description: Bad Request
        '404':
          description: Not Found
    post:
      parameters:
        - name: group 
          in: query
          required: true
          schema:
            type: string
        - name: song
          in: query
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongCreated'
        '500':
          description: Internal Server Error
  /songs/all:
    get:
      parameters:
        - name: group 
          in: query
          required: false
          schema:
            type: string
        - name: song
          in: query
          required: false
          schema:
            type: string
        - name: releaseDate
          in: query
          required: false
          schema:
            type: string
        - name: page 
          in: query
          required: false
          schema:
            type: integer
        - name: size
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllSongs'
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /songs/{id}:
    patch:
      parameters:
        - name: id
          in: path
          description: Song ID
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                songName:
                  type: string
                  example: "Heartbeat"
                groupName:
                  type: string
                  example: "Childish Gambino"
                text:
                  type: string
                  example: "ooirbhirrtih\nnggeriHEARTBEEATgheriovHEARTBEEAThreoin  \n\nthbriuhvtruivr"
                link:
                  type: string
                  example: "https://www.amalgama-lab.com/songs/c/childish_gambino/heartbeat.html"
                releaseDate:
                  type: string
                  example: "16.07.2006"
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    delete:
      parameters:
        - name: id
          in: path
          description: Song ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          
components:
  schemas:
    CoupletsDetails:
      required:
        - totalCount
        - page
        - size
        - text
      type: object
      properties:
        totalCount:
          type: integer
          example: 12
        page:
          type: integer
          example: 2
        size:
          type: integer
          example: 4
        text:
          type: array
          items:
            type: string
          example: ["one", "two", "three"]
    SongCreated:
      required:
      - songName
      - groupName
      - link
      type: object
      properties:
        songName:
          type: string
          example: "Deep end"
        groupName:
          type: string
          example: "Libercio"
        link:
          type: string
          example: https://www.youtube.com/watch?v=Xsp3_a-PMTw
    AllSongs:
      required:
      - totalCount
      - page
      - size
      - songs
      type: object
      properties:
        totalCount:
          type: integer
          example: 12
        page:
          type: integer
          example: 2
        size:
          type: integer
          example: 4
        songs:
          type: array
          items:
            type: object
            required:
              - songName
              - groupName
              - link
              - releaseDate
              - text
            properties:
              songName:
                type: string
                example: "Deep end"
              groupName:
                type: string
                example: "Libercio"
              link:
                type: string
                example: https://www.youtube.com/watch?v=Xsp3_a-PMTw
              releaseDate:
                type: string
                example: 16.07.2006
              text:
                type: array
                items:
                  type: string
                example: ["one", "two", "three"]